refactor: rename types into quantum_numbers
