+++ expertsystem/solving/quantum_numbers.py
