+++ tests/unit/solving/conservation_rules/test_mass.py
@@ -4,7 +4,7 @@ from expertsystem.solving.conservation_rules import (
     MassConservation,
     MassEdgeInput,
 )
-from expertsystem.solving.types import EdgeQuantumNumbers
+from expertsystem.solving.quantum_numbers import EdgeQuantumNumbers
 
 # Currently need to cast to the proper Edge/NodeQuantumNumber type, see
 # https://github.com/ComPWA/expertsystem/issues/255
